Index

bigleaf.ET_to_LE

bigleaf.Esat_from_Tair

bigleaf.Esat_from_Tair_deriv

bigleaf.Esat_slope

bigleaf.Esat_slope

bigleaf.LE_to_ET

bigleaf.LE_to_ET

bigleaf.PPFD_to_Rg

bigleaf.Rg_to_PPFD

bigleaf.Rg_to_PPFD

bigleaf.VPD_to_rH

bigleaf.VPD_to_rH

bigleaf.bigleaf_constants

bigleaf.gC_to_umolCO2

bigleaf.kg_to_mol

bigleaf.kg_to_mol

bigleaf.latent_heat_vaporization

bigleaf.latent_heat_vaporization

bigleaf.mol_to_ms

bigleaf.ms_to_mol

bigleaf.ms_to_mol

bigleaf.umolCO2_to_gC

bigleaf.umolCO2_to_gC

bigleaf.virtual_temp

Autodocs

bigleaf.ET_to_LE — Method

LE_to_ET(LE,Tair)
ET_to_LE(ET,Tair)

Convert evaporative water flux from mass (ET=evapotranspiration) to energy (LE=latent heat flux) units, or vice versa.

Arguments

LE Latent heat flux (W m-2)

ET Evapotranspiration (kg m-2 s-1)

Tair Air temperature (deg C)

Details

The conversions are given by:

$ET = LE/\lambda$

$LE = \lambda ET$

where $\lambda$ is the latent heat of vaporization (J kg-1) as calculated by latent_heat_vaporization.

Examples

# LE of 200 Wm-2 and air temperature of 25degC
LE_to_ET(200,25)

x = linspace(-π, π) # hide
plot(x, f(x), color = "red")
savefig("f-plot.svg"); 

bigleaf.Esat_from_Tair — Method

Esat_slope(Tair; formula, constants) 
Esat_from_Tair(Tair; formula, constants) 
Esat_from_Tair_deriv(Tair; formula, constants)

Saturation Vapor Pressure (Esat) and Slope of the Esat Curve

Arguemtns

Tair: Air temperature (deg C)

formula=Val(:Sonntag_1990): Formula to be used. Either Val(:Sonntag_1990) (Default), Val(:Alduchov_1996, or Val(:Allen_1998).

constants=bigleaf_constants(): Dictionary with entry :Pa2kPa - conversion pascal (Pa) to kilopascal

Details

Esat (kPa) is calculated using the Magnus equation:

Esat = a * exp((b * Tair) / (c + Tair)) / 1000 where the coefficients a, b, c take different values depending on the formula use The default values are from Sonntag 1990 (a=611.2, b=17.62, c=243.12). This versi of the Magnus equation is recommended by the WMO (WMO 2008; p1.4-29). Alternativel parameter values determined by Alduchov & Eskridge 1996 or Allen et al. 1998 can b used (see references). The slope of the Esat curve ($\Delta$) is calculated as the first derivative of the function:

$\Delta = {dEsat \over dTair}$

Value

Esat_from_Tair: Saturation vapor pressure (kPa)

Esat_from_Tair_deriv: Slope of the saturation vapor pressure curve (kPa K-1)

Esat_slope: A tuple of the two above values

References

Sonntag D. 1990: Important new values of the physical constants of 1986, vapor pressure formulations based on the ITS-90 and psychrometric formulae. Zeitschrift fuer Meteorologie 70, 340-344.

World Meteorological Organization 2008: Guide to Meteorological Instruments and Methods of Observation (WMO-No.8). World Meteorological Organization, Geneva. 7th Edition,

Alduchov, O. A. & Eskridge, R. E., 1996: Improved Magnus form approximation of saturation vapor pressure. Journal of Applied Meteorology, 35, 601-609

Allen, R.G., Pereira, L.S., Raes, D., Smith, M., 1998: Crop evapotranspiration - Guidelines for computing crop water requirements - FAO irrigation and drainage paper 56, FAO, Rome.

Esat_from_Tair(20.0)          # Esat in kPa
Esat_from_Tair_deriv(20.0)    # its derivative to temperature in kPa K-1
Esat_slope(20.0)              # both as a tuple

bigleaf.Esat_from_Tair_deriv — Method

Esat_slope(Tair; formula, constants) 
Esat_from_Tair(Tair; formula, constants) 
Esat_from_Tair_deriv(Tair; formula, constants)

Saturation Vapor Pressure (Esat) and Slope of the Esat Curve

Arguemtns

Tair: Air temperature (deg C)

formula=Val(:Sonntag_1990): Formula to be used. Either Val(:Sonntag_1990) (Default), Val(:Alduchov_1996, or Val(:Allen_1998).

constants=bigleaf_constants(): Dictionary with entry :Pa2kPa - conversion pascal (Pa) to kilopascal

Details

Esat (kPa) is calculated using the Magnus equation:

Esat = a * exp((b * Tair) / (c + Tair)) / 1000 where the coefficients a, b, c take different values depending on the formula use The default values are from Sonntag 1990 (a=611.2, b=17.62, c=243.12). This versi of the Magnus equation is recommended by the WMO (WMO 2008; p1.4-29). Alternativel parameter values determined by Alduchov & Eskridge 1996 or Allen et al. 1998 can b used (see references). The slope of the Esat curve ($\Delta$) is calculated as the first derivative of the function:

$\Delta = {dEsat \over dTair}$

Value

Esat_from_Tair: Saturation vapor pressure (kPa)

Esat_from_Tair_deriv: Slope of the saturation vapor pressure curve (kPa K-1)

Esat_slope: A tuple of the two above values

References

Sonntag D. 1990: Important new values of the physical constants of 1986, vapor pressure formulations based on the ITS-90 and psychrometric formulae. Zeitschrift fuer Meteorologie 70, 340-344.

World Meteorological Organization 2008: Guide to Meteorological Instruments and Methods of Observation (WMO-No.8). World Meteorological Organization, Geneva. 7th Edition,

Alduchov, O. A. & Eskridge, R. E., 1996: Improved Magnus form approximation of saturation vapor pressure. Journal of Applied Meteorology, 35, 601-609

Allen, R.G., Pereira, L.S., Raes, D., Smith, M., 1998: Crop evapotranspiration - Guidelines for computing crop water requirements - FAO irrigation and drainage paper 56, FAO, Rome.

Esat_from_Tair(20.0)          # Esat in kPa
Esat_from_Tair_deriv(20.0)    # its derivative to temperature in kPa K-1
Esat_slope(20.0)              # both as a tuple

bigleaf.Esat_slope — Method

Esat_slope(Tair; formula, constants) 
Esat_from_Tair(Tair; formula, constants) 
Esat_from_Tair_deriv(Tair; formula, constants)

Saturation Vapor Pressure (Esat) and Slope of the Esat Curve

Arguemtns

Tair: Air temperature (deg C)

formula=Val(:Sonntag_1990): Formula to be used. Either Val(:Sonntag_1990) (Default), Val(:Alduchov_1996, or Val(:Allen_1998).

constants=bigleaf_constants(): Dictionary with entry :Pa2kPa - conversion pascal (Pa) to kilopascal

Details

Esat (kPa) is calculated using the Magnus equation:

Esat = a * exp((b * Tair) / (c + Tair)) / 1000 where the coefficients a, b, c take different values depending on the formula use The default values are from Sonntag 1990 (a=611.2, b=17.62, c=243.12). This versi of the Magnus equation is recommended by the WMO (WMO 2008; p1.4-29). Alternativel parameter values determined by Alduchov & Eskridge 1996 or Allen et al. 1998 can b used (see references). The slope of the Esat curve ($\Delta$) is calculated as the first derivative of the function:

$\Delta = {dEsat \over dTair}$

Value

Esat_from_Tair: Saturation vapor pressure (kPa)

Esat_from_Tair_deriv: Slope of the saturation vapor pressure curve (kPa K-1)

Esat_slope: A tuple of the two above values

References

Sonntag D. 1990: Important new values of the physical constants of 1986, vapor pressure formulations based on the ITS-90 and psychrometric formulae. Zeitschrift fuer Meteorologie 70, 340-344.

World Meteorological Organization 2008: Guide to Meteorological Instruments and Methods of Observation (WMO-No.8). World Meteorological Organization, Geneva. 7th Edition,

Alduchov, O. A. & Eskridge, R. E., 1996: Improved Magnus form approximation of saturation vapor pressure. Journal of Applied Meteorology, 35, 601-609

Allen, R.G., Pereira, L.S., Raes, D., Smith, M., 1998: Crop evapotranspiration - Guidelines for computing crop water requirements - FAO irrigation and drainage paper 56, FAO, Rome.

Esat_from_Tair(20.0)          # Esat in kPa
Esat_from_Tair_deriv(20.0)    # its derivative to temperature in kPa K-1
Esat_slope(20.0)              # both as a tuple

bigleaf.LE_to_ET — Method

LE_to_ET(LE,Tair)
ET_to_LE(ET,Tair)

Convert evaporative water flux from mass (ET=evapotranspiration) to energy (LE=latent heat flux) units, or vice versa.

Arguments

LE Latent heat flux (W m-2)

ET Evapotranspiration (kg m-2 s-1)

Tair Air temperature (deg C)

Details

The conversions are given by:

$ET = LE/\lambda$

$LE = \lambda ET$

where $\lambda$ is the latent heat of vaporization (J kg-1) as calculated by latent_heat_vaporization.

Examples

# LE of 200 Wm-2 and air temperature of 25degC
LE_to_ET(200,25)

x = linspace(-π, π) # hide
plot(x, f(x), color = "red")
savefig("f-plot.svg"); 

bigleaf.PPFD_to_Rg — Function

Rg_to_PPFD(Rg,J_to_mol=4.6,frac_PAR=0.5)
PPFD_to_Rg(PPFD,J_to_mol=4.6,frac_PAR=0.5)

Conversions between Global Radiation (W m-2) and Photosynthetic Photon Flux Density (umol m-2 s-1)

Arguments

Rg: Global radiation = incoming short-wave radiation at the surface (W m-2)

PPFD: Photosynthetic photon flux density (umol m-2 s-1)

Jtomol: Conversion factor from J m-2 s-1 (= W m-2) to umol (quanta) m-2 s-1

frac_PAR: Fraction of incoming solar irradiance that is photosynthetical- active radiation (PAR); defaults to 0.5

Details

The conversion is given by:

$PPFD = Rg * frac_PAR * J_to_mol$

by default, the combined conversion factor (frac_PAR * J_to_mol) is 2.3

Examples

# convert a measured incoming short-wave radiation of 500 Wm-2 to
# PPFD in umol m-2 s-1 and backwards
Rg_to_PPFD(500)
PPFD_to_Rg(1150)

bigleaf.Rg_to_PPFD — Function

Rg_to_PPFD(Rg,J_to_mol=4.6,frac_PAR=0.5)
PPFD_to_Rg(PPFD,J_to_mol=4.6,frac_PAR=0.5)

Conversions between Global Radiation (W m-2) and Photosynthetic Photon Flux Density (umol m-2 s-1)

Arguments

Rg: Global radiation = incoming short-wave radiation at the surface (W m-2)

PPFD: Photosynthetic photon flux density (umol m-2 s-1)

Jtomol: Conversion factor from J m-2 s-1 (= W m-2) to umol (quanta) m-2 s-1

frac_PAR: Fraction of incoming solar irradiance that is photosynthetical- active radiation (PAR); defaults to 0.5

Details

The conversion is given by:

$PPFD = Rg * frac_PAR * J_to_mol$

by default, the combined conversion factor (frac_PAR * J_to_mol) is 2.3

Examples

# convert a measured incoming short-wave radiation of 500 Wm-2 to
# PPFD in umol m-2 s-1 and backwards
Rg_to_PPFD(500)
PPFD_to_Rg(1150)

bigleaf.VPD_to_rH — Method

VPD_to_rH(VPD,Tair; ...)
H_to_VPD(rH,Tair; ...)
e_to_rH(e,Tair; ...)
VPD_to_e(VPD,Tair; ...)
e_to_VPD(e,Tair; ...)
e_to_q(e,pressure; ...)
q_to_e(q,pressure; ...)
q_to_VPD(q,Tair,pressure; ...)
VPD_to_q(VPD,Tair,pressure; ...)

Conversion between vapor pressure (e), vapor pressure deficit (VPD), specific humidity (q), and relative humidity (rH).

Arguments

Tair: Air temperature (deg C)

pressure: Atmospheric pressure (kPa)

e: Vapor pressure (kPa)

q: Specific humidity (kg kg-1)

VPD: Vapor pressure deficit (kPa)

rH: Relative humidity (-)

All functions accept the optional arguemtns:

Esat_formula: formula used in Esat_from_Tair

constants: dictionary from bigleaf_constants with entries eps and Pa2kPa

Rreferences

Foken, T, 2008: Micrometeorology_ Springer, Berlin, Germany.

bigleaf.bigleaf_constants — Method

bigleaf_constants(;...)

Constants Used in the bigleaf.jl Package

Return a Dictionary Dict{Symbol, Real}. Default values can be overridden by the named arguments:

cp Specific heat of air for constant pressure (J K-1 kg-1)

Rgas Universal gas constant (J mol-1 K-1)

Rv Gas constant of water vapor (J kg-1 K-1) (Stull 1988 p.641)

Rd Gas constant of dry air (J kg-1 K-1) (Foken p. 245)

Md Molar mass of dry air (kg mol-1)

Mw Molar mass of water vapor (kg mol-1)

eps Ratio of the molecular weight of water vapor to dry air (=Mw/Md)

g Gravitational acceleration (m s-2)

solar_constant Solar constant (W m-2)

pressure0 Reference atmospheric pressure at sea level (Pa)

Tair0 Reference air temperature (K)

k von Karman constant

Cmol Molar mass of carbon (kg mol-1)

Omol Molar mass of oxygen (kg mol-1)

H2Omol Molar mass of water (kg mol-1)

sigma Stefan-Boltzmann constant (W m-2 K-4)

Pr Prandtl number

Sc_CO2 Schmidt number for CO2

Kelvin Conversion degree Celsius to Kelvin

DwDc Ratio of the molecular diffusivities for water vapor and CO2

days2seconds Seconds per day

kPa2Pa Conversion kilopascal (kPa) to pascal (Pa)

Pa2kPa Conversion pascal (Pa) to kilopascal (kPa)

umol2mol Conversion micromole (umol) to mole (mol)

mol2umol Conversion mole (mol) to micromole (umol)

kg2g Conversion kilogram (kg) to gram (g)

g2kg Conversion gram (g) to kilogram (kg)

kJ2J Conversion kilojoule (kJ) to joule (J)

J2kJ Conversion joule (J) to kilojoule (kJ)

se_median Conversion standard error (SE) of the mean to SE of the median

frac2percent Conversion between fraction and percent

bigleaf.gC_to_umolCO2 — Method

umolCO2_to_gC(CO2_flux; constants=bigleaf_constants())
gC_to_umolCO2(C_flux; constants=bigleaf_constants())

Convert CO2 quantities from (umol CO2 m-2 s-1) to (g C m-2 d-1) and vice versa.

Arguments

CO2_flux CO2 flux (umol CO2 m-2 s-1)

C_flux Carbon (C) flux (gC m-2 d-1)

constants: dictionary from bigleaf_constants with entries: Cmol, umol2mol, mol2umol, kg2g, g2kg, says2seconds

Examples

umolCO2_to_gC(20)  # gC m-2 d-1

bigleaf.kg_to_mol — Function

kg_to_mol(mass, molarMass=bigleaf_constants()[:H2Omol])

Conversion between Mass (kg) and Molar Units (mol).

bigleaf.latent_heat_vaporization — Method

latent_heat_vaporization(Tair)

Latent heat of vaporization as a function of air temperature using

$\lambda = (2.501 - 0.00237 \, Tair) 10^6$.

Arguments:

Tair: Air temperature (deg C)

Value

$\lambda$: Latent heat of vaporization (J kg-1)

References

Stull, B, 1988: An Introduction to Boundary Layer Meteorology (p641) Kluwer Academic Publishers, Dordrecht, Netherlands

Foken, T, 2008: Micrometeorology_ Springer, Berlin, Germany

latent_heat_vaporization.(range(5,45,5))        

bigleaf.mol_to_ms — Method

ms_to_mol(G_ms,Tair,pressure; constants=bigleaf_constants())
mol_to_ms(G_mol,Tair,pressure; constants=bigleaf_constants())

Converts conductances from mass (m s-1) to molar units (mol m-2 s-1), or vice versa

Details

The conversions are given by

$G_{mol} = G_{ms} \, pressure / (Rgas Tair)$

$G_{ms} = G_{mol} \, (Rgas Tair) / pressure$

where Tair is in Kelvin and pressure in Pa (converted from kPa internally).

References

Jones, HG 1992_ Plants and microclimate: a quantitative approach to environmental plant physiology_ 2nd Edition, Cambridge University Press, Cambridge 428

ms_to_mol(0.005,25,100)

bigleaf.ms_to_mol — Method

ms_to_mol(G_ms,Tair,pressure; constants=bigleaf_constants())
mol_to_ms(G_mol,Tair,pressure; constants=bigleaf_constants())

Converts conductances from mass (m s-1) to molar units (mol m-2 s-1), or vice versa

Details

The conversions are given by

$G_{mol} = G_{ms} \, pressure / (Rgas Tair)$

$G_{ms} = G_{mol} \, (Rgas Tair) / pressure$

where Tair is in Kelvin and pressure in Pa (converted from kPa internally).

References

Jones, HG 1992_ Plants and microclimate: a quantitative approach to environmental plant physiology_ 2nd Edition, Cambridge University Press, Cambridge 428

ms_to_mol(0.005,25,100)

bigleaf.umolCO2_to_gC — Method

umolCO2_to_gC(CO2_flux; constants=bigleaf_constants())
gC_to_umolCO2(C_flux; constants=bigleaf_constants())

Convert CO2 quantities from (umol CO2 m-2 s-1) to (g C m-2 d-1) and vice versa.

Arguments

CO2_flux CO2 flux (umol CO2 m-2 s-1)

C_flux Carbon (C) flux (gC m-2 d-1)

constants: dictionary from bigleaf_constants with entries: Cmol, umol2mol, mol2umol, kg2g, g2kg, says2seconds

Examples

umolCO2_to_gC(20)  # gC m-2 d-1

bigleaf.virtual_temp — Method

virtual_temp(Tair, pressure, VPD; Esat_formula, constants)

Virtual temperature, defined as the temperature at which dry air would have the same density as moist air at its actual temperature.

Arguments

Tair: Air temperature (deg C)

pressure: Atmospheric pressure (kPa)

VPD: Vapor pressure deficit (kPa)

Esat_formula: formula used in Esat_from_Tair

constants Dictionary with entries

:Kelvin - conversion degree Celsius to Kelvin

:eps - ratio of the molecular weight of water vapor to dry air (-)

Details

The virtual temperature is given by:

Tv = Tair / (1 - (1 - eps) e/pressure)

where Tair is in Kelvin (converted internally)_ Likewise, VPD is converted to actual vapor pressure (e in kPa) with VPD_to_e internally.

Value

virtual temperature (deg C)

References

Monteith JL, Unsworth MH, 2008: Principles of Environmental Physics. 3rd edition_ Academic Press, London.

Examples

virtual_temp(25,100,1.5)