Global radiation

Bigleaf.calc_sun_position_hor — Function

calc_sun_position_hor(datetime::DateTime, lat, long)

Compute the Sun position at given time and observer coordinates in horizontal coordinates.

Arguments:

datetime: time given in UTC.

lat, long: latitude and longitude in degree

Value

SLVector: sun position with entries

altitude: angle above the horizon [rad].

azimuth: angle ange the horizon plane eastwards of north [rad]

hourangle: [rad] as output by AstroLib.eq2hor Seems to represent time [day/2pi] after solar noon. Value at local timezone noon provdes (local time - solar time).

Bigleaf.calc_sun_position_MOD — Function

calc_sun_position_MOD(JD::Number)

Compute the Sun position at the Julian Day JD.

Results are represented in the Mean Equinox of Date (MOD), i.e. accounting for precision but not for nutation and smaller pertubation of the x axes, in Spherical coordinates. The algorithm was adapted from [Vallado 2013, p. 277-279].

Arguments:

JD: time given as Julian Day .

Value

SLVector: sun position where

Ecliptic coordinates (1:3)

λ: Ecliptic longitude of the Sun [rad].

β: Ecliptic latitude of the Sun [rad] is assumed 0.

r: Distance of the Sun from Earth [m].

Equatorial coordinate (4:5)

α: ascention [rad]

δ: declination [rad]

ϵ: Obliquity of the ecliptic [rad].

References

Vallado, D. A (2013). Fundamentals of Astrodynamics and Applications. Microcosm Press, Hawthorne, CA.