Meteorological variables

bigleaf.latent_heat_vaporization — Function

latent_heat_vaporization(Tair)

Latent heat of vaporization as a function of air temperature using

$\lambda = (2.501 - 0.00237 \, Tair) 10^6$.

Arguments:

Tair: Air temperature (deg C)

Value

$\lambda$: Latent heat of vaporization (J kg-1)

References

Stull, B, 1988: An Introduction to Boundary Layer Meteorology (p641) Kluwer Academic Publishers, Dordrecht, Netherlands

Foken, T, 2008: Micrometeorology_ Springer, Berlin, Germany

latent_heat_vaporization.(range(5,45,5))        

Missing docstring.

Missing docstring for virtual_temp. Check Documenter's build log for details.

bigleaf.Esat_slope — Function

Esat_slope(Tair; formula, constants) 
Esat_from_Tair(Tair; formula, constants) 
Esat_from_Tair_deriv(Tair; formula, constants)

Saturation Vapor Pressure (Esat) and Slope of the Esat Curve

Arguemtns

Tair: Air temperature (deg C)

formula=Val(:Sonntag_1990): Formula to be used. Either Val(:Sonntag_1990) (Default), Val(:Alduchov_1996, or Val(:Allen_1998).

constants=bigleaf_constants(): Dictionary with entry :Pa2kPa - conversion pascal (Pa) to kilopascal

Details

Esat (kPa) is calculated using the Magnus equation:

Esat = a * exp((b * Tair) / (c + Tair)) / 1000 where the coefficients a, b, c take different values depending on the formula use The default values are from Sonntag 1990 (a=611.2, b=17.62, c=243.12). This versi of the Magnus equation is recommended by the WMO (WMO 2008; p1.4-29). Alternativel parameter values determined by Alduchov & Eskridge 1996 or Allen et al. 1998 can b used (see references). The slope of the Esat curve ($\Delta$) is calculated as the first derivative of the function:

$\Delta = {dEsat \over dTair}$

Value

Esat_from_Tair: Saturation vapor pressure (kPa)

Esat_from_Tair_deriv: Slope of the saturation vapor pressure curve (kPa K-1)

Esat_slope: A tuple of the two above values

References

Sonntag D. 1990: Important new values of the physical constants of 1986, vapor pressure formulations based on the ITS-90 and psychrometric formulae. Zeitschrift fuer Meteorologie 70, 340-344.

World Meteorological Organization 2008: Guide to Meteorological Instruments and Methods of Observation (WMO-No.8). World Meteorological Organization, Geneva. 7th Edition,

Alduchov, O. A. & Eskridge, R. E., 1996: Improved Magnus form approximation of saturation vapor pressure. Journal of Applied Meteorology, 35, 601-609

Allen, R.G., Pereira, L.S., Raes, D., Smith, M., 1998: Crop evapotranspiration - Guidelines for computing crop water requirements - FAO irrigation and drainage paper 56, FAO, Rome.

Esat_from_Tair(20.0)          # Esat in kPa
Esat_from_Tair_deriv(20.0)    # its derivative to temperature in kPa K-1
Esat_slope(20.0)              # both as a tuple